DOM-2537 refactored after review
